+++ src/KiotaClientFactory.cs
@@ -55,6 +55,7 @@ namespace Microsoft.Kiota.Http.HttpClientLibrary
 
             return new List<DelegatingHandler>
             {
+                //add the default middlewares as they are ready, and add them to the list below as well
                 //add the default middlewares as they are ready
                 
                 optionsForHandlers.OfType<UriReplacementHandlerOption>().FirstOrDefault() is UriReplacementHandlerOption uriReplacementOption
@@ -82,6 +83,25 @@ namespace Microsoft.Kiota.Http.HttpClientLibrary
                 : new HeadersInspectionHandler(),
             };
         }
+
+        /// <summary>
+        /// Gets the default handler types.
+        /// </summary>
+        /// <returns>A list of all the default handlers</returns>
+        /// <remarks>Order matters</remarks>
+        public static IList<System.Type> GetDefaultHandlerTypes()
+        {
+            return new List<System.Type>
+            {
+                typeof(UriReplacementHandler<UriReplacementHandlerOption>),
+                typeof(RetryHandler),
+                typeof(RedirectHandler),
+                typeof(ParametersNameDecodingHandler),
+                typeof(UserAgentHandler),
+                typeof(HeadersInspectionHandler),
+            };
+        }
+
         /// <summary>
         /// Creates a <see cref="DelegatingHandler"/> to use for the <see cref="HttpClient" /> from the provided <see cref="DelegatingHandler"/> instances. Order matters.
         /// </summary>
@@ -103,7 +123,7 @@ namespace Microsoft.Kiota.Http.HttpClientLibrary
                 }
             }
             if(finalHandler != null)
-                handlers[handlers.Length-1].InnerHandler = finalHandler;
+                handlers[handlers.Length - 1].InnerHandler = finalHandler;
             return handlers[0];//first
         }
         /// <summary>
@@ -113,7 +133,7 @@ namespace Microsoft.Kiota.Http.HttpClientLibrary
         /// <returns>The created <see cref="DelegatingHandler"/>.</returns>
         public static DelegatingHandler? ChainHandlersCollectionAndGetFirstLink(params DelegatingHandler[] handlers)
         {
-            return ChainHandlersCollectionAndGetFirstLink(null,handlers);
+            return ChainHandlersCollectionAndGetFirstLink(null, handlers);
         }
         /// <summary>
         /// Gets a default Http Client handler with the appropriate proxy configurations
@@ -126,7 +146,9 @@ namespace Microsoft.Kiota.Http.HttpClientLibrary
             // If custom proxy is passed, the WindowsProxyUsePolicy will need updating
             // https://github.com/dotnet/runtime/blob/main/src/libraries/System.Net.Http.WinHttpHandler/src/System/Net/Http/WinHttpHandler.cs#L575
             var proxyPolicy = proxy != null ? WindowsProxyUsePolicy.UseCustomProxy : WindowsProxyUsePolicy.UseWinHttpProxy;
-            return new WinHttpHandler { Proxy = proxy, AutomaticDecompression = DecompressionMethods.None , WindowsProxyUsePolicy = proxyPolicy, SendTimeout = System.Threading.Timeout.InfiniteTimeSpan, ReceiveDataTimeout = System.Threading.Timeout.InfiniteTimeSpan, ReceiveHeadersTimeout = System.Threading.Timeout.InfiniteTimeSpan };
+            return new WinHttpHandler { Proxy = proxy, AutomaticDecompression = DecompressionMethods.None, WindowsProxyUsePolicy = proxyPolicy, SendTimeout = System.Threading.Timeout.InfiniteTimeSpan, ReceiveDataTimeout = System.Threading.Timeout.InfiniteTimeSpan, ReceiveHeadersTimeout = System.Threading.Timeout.InfiniteTimeSpan, EnableMultipleHttp2Connections = true };
+#elif NET5_0_OR_GREATER
+            return new SocketsHttpHandler { Proxy = proxy, AllowAutoRedirect = false, EnableMultipleHttp2Connections = true };
 #else
             return new HttpClientHandler { Proxy = proxy, AllowAutoRedirect = false };
 #endif